Fix MaxPages type to uint16 Add Unused array to InitOut

+++ ops.go
@@ -44,5 +44,5 @@ type initOp struct {
 	MaxReadahead  uint32
 	MaxBackground uint16
 	MaxWrite      uint32
-	MaxPages      uint32
+	MaxPages      uint16
 }
